Better organize code.
