+++ src/DependencyManager.ts
@@ -2,8 +2,7 @@
 // Licensed under the MIT license.
 
 import { QuickPickItem } from "vscode";
-import { IDependency } from "./Interfaces";
-import * as Metadata from "./Metadata";
+import { getAvailableDependencies, IDependency } from "./model";
 import { readFileFromExtensionRoot, writeFileToExtensionRoot } from "./Utils";
 
 const PLACEHOLDER: string = "";
@@ -33,7 +32,7 @@ class DependencyManager {
 
     public async getQuickPickItems(bootVersion: string, options?: { hasLastSelected: boolean }): Promise<Array<QuickPickItem & IDependenciesItem>> {
         if (this.dependencies.length === 0) {
-            await this.initialize(await Metadata.getAvailableDependencies(bootVersion));
+            await this.initialize(await getAvailableDependencies(bootVersion));
         }
         const ret: Array<QuickPickItem & IDependenciesItem> = [];
         if (this.selectedIds.length === 0) {
