+++ src/handler/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { EditStartersHandler } from "./EditStartersHandler";
+import { GenerateProjectHandler } from "./GenerateProjectHandler";
+
+export { EditStartersHandler, GenerateProjectHandler };
