+++ src/model/Interfaces.ts
