+++ src/model/pomxml/BomNode.ts
