+++ src/model/pomxml/DependencyNode.ts
